requirements.txt
requests
dnspython
python-whois
shodan
tabulate